requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
google-colab  # Only needed if running in Google Colab
ipython>=8.0.0  # For Javascript/display functionality 